+++ demo/basic/contextmenu.component.ts
@@ -13,7 +13,7 @@ import { Component } from '@angular/core';
         </small>
       </h3>
       <div class="info">
-        <p><strong>Note:</strong> angular2-data-table does not provide a context menu feature.
+        <p><strong>Note:</strong> ngx-datatable does not provide a context menu feature.
         This demonstrates how you would access the <code>contextmenu</code> event
         to display your own custom context menu.</p>
         <p *ngIf="rawEvent"><strong>Mouse position:</strong> <code>(x: {{rawEvent?.x}}, y: {{rawEvent?.y}})</code></p>