+++ release/components/body/scroller.component.js
@@ -58,7 +58,8 @@ var ScrollerComponent = (function () {
                     template: "\n    <ng-content></ng-content>\n  ",
                     host: {
                         class: 'datatable-scroll'
-                    }
+                    },
+                    changeDetection: core_1.ChangeDetectionStrategy.OnPush
                 },] },
     ];
     /** @nocollapse */