+++ docs/changelog.md
@@ -1,4 +1,13 @@
 # Changelog
+## 10.0.5
+- Style: Material styles closer to spec
+- Bug: Non-sortable header showing pointer cursor
+
+## 10.0.4
+- Perf: Changed everything internal to OnPush
+- Perf: Fixed header cell outlet cd
+- Bug: Fixed sort mutating original rows
+
 ## 10.0.3
 - Perf: Better performance in cells